complete happy path

+++ services/product-catalog-service/src/ProductCatalogService.Application/Common/Dtos/ProductDto.cs
@@ -83,9 +83,9 @@ public class ShippingInfoDto
 {
     public int WeightGrams { get; init; }
 
-    public double Length { get; init; }
+    public int Length { get; init; }
 
-    public double Width { get; init; }
+    public int Width { get; init; }
 
-    public double Height { get; init; }
+    public int Height { get; init; }
 }
