+++ services/product-catalog-service/src/ProductCatalogService.Domain/Entities/Product.cs
@@ -84,9 +84,9 @@ public class ShippingInfo
 
 public class Dimensions
 {
-    public double Length { get; init; }
+    public int Length { get; init; }
 
-    public double Width { get; init; }
+    public int Width { get; init; }
 
-    public double Height { get; init; }
+    public int Height { get; init; }
 }
